chore: migrated the app from the old api to my new spring boot api

+++ api/src/main/kotlin/dev/logickoder/qrpay/transaction/TransactionEntity.kt
@@ -50,6 +50,6 @@ internal fun TransactionEntity.toTransaction() = Transaction(
     id = id,
     type = type,
     description = description,
-    amount = amount,
+    amount = amount.toFloat(),
     time = time,
 )
